+++ SQL/New_patches/2021-07-29_modify_center_name_in_mri_protocol.sql
@@ -1,0 +1,3 @@
+ALTER TABLE mri_protocol ADD `CenterID` integer unsigned DEFAULT NULL AFTER `Center_name`;
+ALTER TABLE mri_protocol ADD FOREIGN KEY (`CenterID`) REFERENCES psc(`CenterID`);
+ALTER TABLE mri_protocol DROP COLUMN `Center_name`;